simplify Job interface
We were instantiating Job objects with two
arguments: job directory path and script name
(or path relative to job directory).

This extra complexity is not required now. If every
job directory is the directory containing the job script
this can be derived from the full path to the script.

So instead of

    Job.new(path: jobdir_path, script: script_name)

We now do

    Job.new(script: script_path)

If we need more flexibility later, i.e. we want PBS_O_WORKDIR
to not be in the same directory as the script, we can add support then.