split SimpleJob into two sub-modules
Move the methods from SimpleJob to SimpleJob::Statusable
and SimpleJob::Submittable. If including SimpleJob directly
this in turn includes both Submittable and Statusable, so this
is backwards compatible with previous changes. But now you can
include the modules individually, so in a Rails app one model
(a workflow model) could include Submittable, and another model
(a batch job model) could include Statusable.