wasn't excluding .git / added .gitignore filter
It was only excluding '.git' subdirectory located inside of a '.svn' directory. Not sure if that was intended, but assumed it wanted to exclude both '.git' directory and '.svn' directory. Also added a filter to exclude files inside '.gitignore' to ignore developmental files not intended to be used in production environment.