add workflow and job setter on statusable
statusable gets a job= where you can set OSC::Machete::Job object
and set the properties from this; a shortcut that is used by
OSC::Machete::SimpleJob::Workflow and could be used elsewhere

OSC::Machete::SimpleJob::Workflow is added, and this actually
is a first Rails plugin, which adds the method "has_machete_workflow_of"
method to ActiveRecord::Base if it is defined - so this technically still
optionally works in Rails but "could" work outside of Rails (with some work)

Calling that method, specifying the relation for the jobs that the model contains
in a has_many relation, results in adding a bunch of methods that refer to this
relation to save jobs after submitting and to check the status of jobs.

Passes all tests on CFS, but need to test because CFS doesn't have many tests.