sync workflow and statusable interfaces
add Workflow#running? and Statusable#running_queued_or_hold?
to make the status interfaces match

this way whether you are interacting with a Workflow instance
or a Statusable instance you can ask the same status methods

this is a temporary solution that will eventually be replaced
by a refactor to using a value object for status #32