fix problems with status methods
Deprecated Statusable#running? which is a poor name.
Statusable#active? is the replacement.

Deprecated Statusable#running_queued_or_hold?
Statusable#active? is the replacement.

Update Workflow to match this.

Fixed Statusable status methods that do comparisons to status values
in string format by converting the status to a string in case that
status is a symbol instead of a string. (temporary solution)