create status value object for jobs
Status#inspect is used in place of "status_human_readable" because
by definition the purpose of inspect is:
"Returns a string containing a human-readable representation of obj."

Internally stores char representation of status
Comparison to other status value's or strings that have the same char or
symbols that have the same char all work.

Status#to_s returns the char as a string which can be used for serialization
You create a new status object by passing the char into the initializer

If you can add statuses together the order of precendence specifies a single
status that is the result. Precendence is determined by the order of the items
in the hash.

Might consider renaming this JobStatus or OSC::Machete::Job::Status (inner class?).