coerce to string prior to comparison
When doing status comparisions elsewhere in this file
we used #to_s first so we always have a String

This is a result of first attempting to use symbols everwhere
but running into problems after those were stored in the database as
strings.

This and will be addressed by using a StatusValue which will handle
comparisons internally. Then we can do something like:

    if current_status.completed?