Don't set -A header if not valid project
Set the -A header only if account_string is passed in as an argument OR
if the default_account_string is a valid project.

If account_string is passed in, we don't validate it - we trust the
one invoking the method.

This is so that users with appl as the primary group submit without -A
being specified.